# Backend/src/main/java/de/sophieherstein/chat_app_backend/user/dto/UpdatePasswordRequest.java
package de.sophieherstein.chat_app_backend.user.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record UpdatePasswordRequest(
        @NotBlank
        String currentPassword,

        @NotBlank
        @Size(min = 8, max = 100)
        String newPassword
) {
}