# Backend/src/main/java/de/sophieherstein/chat_app_backend/user/dto/UpdateUsernameRequest.java
package de.sophieherstein.chat_app_backend.user.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record UpdateUsernameRequest(
        @NotBlank
        @Size(min = 3, max = 50)
        String username
) {
}