# Frontend/src/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://localhost:8080", "description": "Generated server url" }],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["authentication-controller"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "multipart/form-data": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["authentication-controller"],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          }
        }
      }
    },
    "/api/users/me/username": {
      "patch": {
        "tags": ["user-controller"],
        "operationId": "updateUsername",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUsernameRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
            }
          }
        }
      }
    },
    "/api/users/me/profile-image": {
      "patch": {
        "tags": ["user-controller"],
        "operationId": "updateProfileImage",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": { "profileImage": { "type": "string", "format": "binary" } },
                "required": ["profileImage"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
            }
          }
        }
      }
    },
    "/api/users/me/password": {
      "patch": {
        "tags": ["user-controller"],
        "operationId": "updatePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePasswordRequest" }
            }
          },
          "required": true
        },
        "responses": { "204": { "description": "No Content" } }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": ["user-controller"],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
            }
          }
        }
      }
    },
    "/api/auth/username-available": {
      "get": {
        "tags": ["authentication-controller"],
        "operationId": "isUsernameAvailable",
        "parameters": [
          { "name": "username", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsernameAvailabilityResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "maxLength": 50, "minLength": 3 },
          "password": { "type": "string", "maxLength": 100, "minLength": 8 },
          "profileImage": { "type": "string", "format": "binary" }
        },
        "required": ["password", "username"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "tokenType": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserResponse" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "profileImageUrl": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "minLength": 1 },
          "password": { "type": "string", "minLength": 1 }
        },
        "required": ["password", "username"]
      },
      "UpdateUsernameRequest": {
        "type": "object",
        "properties": { "username": { "type": "string", "maxLength": 50, "minLength": 3 } },
        "required": ["username"]
      },
      "UpdatePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": { "type": "string", "minLength": 1 },
          "newPassword": { "type": "string", "maxLength": 100, "minLength": 8 }
        },
        "required": ["currentPassword", "newPassword"]
      },
      "UsernameAvailabilityResponse": {
        "type": "object",
        "properties": { "available": { "type": "boolean" } }
      }
    }
  }
}